Consolidate buttons to DefaultButton.tsx
